+++ relay-server/src/service.rs
@@ -96,8 +96,8 @@ pub struct ServiceState {
 
 impl ServiceState {
     /// Starts all services and returns addresses to all of them.
-    pub fn start(config: Arc<Config>, runtimes: &Runtimes) -> Result<Self> {
-        let upstream_relay = UpstreamRelayService::new(config.clone()).start_in(&runtimes.upstream);
+    pub fn start(config: Arc<Config>) -> Result<Self> {
+        let upstream_relay = UpstreamRelayService::new(config.clone()).start();
         let test_store = TestStoreService::new(config.clone()).start();
 
         let redis_pool = match config.redis() {
@@ -117,9 +117,8 @@ impl ServiceState {
             upstream_relay.clone(),
             processor.clone(),
         )?
-        .start_in(&runtimes.outcome);
-        let outcome_aggregator =
-            OutcomeAggregator::new(&config, outcome_producer.clone()).start_in(&runtimes.outcome);
+        .start();
+        let outcome_aggregator = OutcomeAggregator::new(&config, outcome_producer.clone()).start();
 
         let global_config = GlobalConfigService::new(config.clone(), upstream_relay.clone());
         let global_config_handle = global_config.handle();
@@ -135,7 +134,7 @@ impl ServiceState {
             config.secondary_aggregator_configs().clone(),
             Some(project_cache.clone().recipient()),
         )
-        .start_in(&runtimes.aggregator);
+        .start();
 
         let metric_stats = MetricStats::new(
             config.clone(),
@@ -144,18 +143,18 @@ impl ServiceState {
         );
 
         #[cfg(feature = "processing")]
-        let store = match runtimes.store {
-            Some(ref rt) => Some(
+        let store = config
+            .processing_enabled()
+            .then(|| {
                 StoreService::create(
                     config.clone(),
                     global_config_handle.clone(),
                     outcome_aggregator.clone(),
                     metric_stats.clone(),
-                )?
-                .start_in(rt),
-            ),
-            None => None,
-        };
+                )
+                .map(|s| s.start())
+            })
+            .transpose()?;
 
         let cogs = CogsService::new(
             &config,
@@ -197,7 +196,6 @@ impl ServiceState {
             upstream_relay.clone(),
             global_config.clone(),
         );
-        let guard = runtimes.project.enter();
         ProjectCacheService::new(
             config.clone(),
             buffer_guard.clone(),
@@ -206,7 +204,6 @@ impl ServiceState {
             redis_pool,
         )
         .spawn_handler(project_cache_rx);
-        drop(guard);
 
         let health_check = HealthCheckService::new(
             config.clone(),
@@ -303,34 +300,6 @@ impl ServiceState {
     /// Returns the address of the [`OutcomeProducer`] service.
     pub fn outcome_aggregator(&self) -> &Addr<TrackOutcome> {
         &self.inner.registry.outcome_aggregator
-    }
-}
-
-/// Contains secondary service runtimes.
-#[derive(Debug)]
-pub struct Runtimes {
-    upstream: Runtime,
-    project: Runtime,
-    aggregator: Runtime,
-    outcome: Runtime,
-    #[cfg(feature = "processing")]
-    store: Option<Runtime>,
-}
-
-impl Runtimes {
-    /// Creates the secondary runtimes required by services.
-    #[allow(unused_variables)]
-    pub fn new(config: &Config) -> Self {
-        Self {
-            upstream: create_runtime("upstream-rt", 2),
-            project: create_runtime("project-rt", 2),
-            aggregator: create_runtime("aggregator-rt", 2),
-            outcome: create_runtime("outcome-rt", 2),
-            #[cfg(feature = "processing")]
-            store: config
-                .processing_enabled()
-                .then(|| create_runtime("store-rt", 1)),
-        }
     }
 }